fix unwanted dtype promotion

+++ netket/experimental/operator/_fermion_operator_2nd_jax.py
@@ -112,7 +112,9 @@ def apply_term(x, w, sites, daggers):
     d = x_at_i != daggers[None, :, None]
     xi = x_at_i[..., jnp.arange(len(sites), dtype=np.uint32), sites]
     d = xi != daggers[None]
-    w_final = w * d.prod(axis=-1) * sgn
+    # here we cast, for the case when x is float64 but weights are float32
+    # and jax would promote the result to float64
+    w_final = w * (d.prod(axis=-1) * sgn).astype(w.dtype)
     return x_final, w_final
